DataForge/backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pydantic
python-multipart
typer
jinja2
rapidfuzz
google-generativeai
kaggle>=1.6.17
datasets
